requirements.txt
# ============================================================
# Pasanaku API - Dependencies
# FastAPI + SQLAlchemy 2.x + PostgreSQL
# ============================================================

# --- Web Framework ---
fastapi==0.115.6
uvicorn[standard]==0.32.1

# --- Database ---
sqlalchemy==2.0.36
alembic==1.14.0
asyncpg==0.30.0

# --- Auth & Security ---
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.2.0
python-multipart==0.0.19

# --- Config & Validation ---
pydantic==2.10.3
pydantic-settings==2.7.0
python-dotenv==1.0.1

# --- HTTP Client ---
httpx==0.28.1

# --- Testing ---
pytest==8.3.4
pytest-asyncio==0.24.0
pytest-cov==6.0.0
httpx==0.28.1

# --- Dev Tools ---
ruff==0.8.4
mypy==1.13.0
